gitlab-ci: Define our own container for ci jobs.
For most projects we use the docker.io/library/gcc:9 base image,
which is based on debian. The build script runs as root inside
that container, which means we can use `apt` to install dependencies.

However, the flac tests fail if they're run as root, so we need
to define a non-privileged user to run the actual build as, and
that means we need a custom base image which provides the necessary
dependencies.

Experimental base image derived from Alpine Linux, which is about
half the size of the equivalent debian-slim container image.